HolderUIをファイル独立、UIComponent継承、デバッグコード移植
